Update to allow relative specification of toolboxRoot in record file.
When this is done, the files und up at toolboxRoot relative to the directory one level up from the toolboxes directory.  The reason I wanted this is because I wanted to allow one project to call routines in another project, that itself has a life of its own.  The other project may already be in the projects folder, or it may not be.  This allows us both to have it fetched but also to use it if it is already there.

This is a bit klugy.  It will break code that is specifying a relative toolboxRoot path, but that would previously have been a very bad idea because where things ended up would be relative to an uncontrolled pwd.

A better way to do what I want might be to be able to have a “project” record of some sort, but that is beyond me right now.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>