Cast record.subfolder to cellstr
For a record to be valid JSON, the value of `subrecord` must be either a single string in double-quotes, or a list of strings where each string is in double-quotes and the list is defined by square brackets. MATLAB interprets both of these as type `string array`
We cast to type `cell array of strings` using `cellstr`, which works as expected for `string array`, but is also robust in that it will work for character-arrays and cell arrays that contain character-arrays, strings, or mixes.
The only thing it will not cast, is a cell array of string arrays.