Support for daily, weekly, monthly updates
The update timestamp for each toolbox is saved to the Matlab prefs (key: ToolboxToolbox.LastUpdates.<toolboxname>). If one of the new options is specified, the current date is compared against the timestamp from the prefs and the update is skipped if the specified timespan did not elapse yet. This allows to find a good compromise between speed and up-to-date dependencies if you have tbUse() commands in your scripts. Needs at least Matlab 2014b